Enhance type safety and suppress unused parameter warnings
- Updated function signatures across multiple files to include (void) for unused parameters, improving clarity and suppressing compiler warnings.
- Added explicit casts to size_t in array length checks to ensure type safety.
- Changed function visibility from public to static where appropriate to limit scope and improve encapsulation.

+++ src/client/cl_cgame.c
@@ -812,7 +812,7 @@ static intptr_t QDECL CL_DllSyscall( intptr_t arg, ... ) {
 
 	args[0] = arg;
 	va_start( ap, arg );
-	for (i = 1; i < ARRAY_LEN( args ); i++ )
+	for (i = 1; (size_t)i < ARRAY_LEN( args ); i++ )
 		args[ i ] = va_arg( ap, intptr_t );
 	va_end( ap );
